✨  cache结构重新设计，将packageDependencies也进行缓存，用来匹配没有版本的库，项目代码重构
