+++ packages/vite-add-cdn-script/lib/config/index.ts
@@ -1,0 +1,1 @@
+export const libName = "vite-add-cdn-script";
