+++ packages/vite-add-cdn-script/lib/types.ts
@@ -1,1 +1,19 @@
-export type PropertyCdn = "jsdelivr" | "unpkg" | "bootcdn" | "cdnjs";
+export enum cdnTypes {
+  jsdelivr = "jsdelivr",
+  unpkg = "unpkg",
+  bootcdn = "bootcdn",
+  cdnjs = "cdnjs",
+}
+
+export type PropertyCdn = `${cdnTypes}`;
+
+export enum EEnforce {
+  PRE = "pre",
+  POST = "post",
+}
+
+export interface IOptions {
+  customScript?: { [key: string]: string };
+  retryTimes?: number;
+  defaultCdns?: PropertyCdn[];
+}
