+++ packages/vite-add-cdn-script/lib/utils/cache.ts
@@ -4,8 +4,21 @@ import fs from "fs";
 import { PropertyCdn } from "../types";
 
 export type CacheFileType = {
-  [packageName: string]: {
-    [version: string]: CacheCellType[];
+  packageDependencies: {
+    [packageName: string]: {
+      // url为对应搜索package.json的url
+      [identification: string]: Dependencies;
+    };
+  };
+  cdnsUrl: {
+    [packageName: string]: {
+      [version: string]: CacheCellType[];
+    };
+  };
+};
+export type Dependencies = {
+  dependencies: {
+    [key: string]: string;
   };
 };
 export enum CdnErrorType {
@@ -29,11 +42,14 @@ export type CacheCellType = successCacheCellType | failCacheCellType;
  * 本地缓存控制类
  */
 class CdnCache {
-  private cdnCache: CacheFileType = {};
+  private cdnCache: CacheFileType = {
+    packageDependencies: {},
+    cdnsUrl: {},
+  };
   private cdnCachePath: string = "";
   constructor() {
     // cdn缓存文件
-    this.cdnCachePath = path.resolve(process.cwd(), "./.cdn-cache-v2.json");
+    this.cdnCachePath = path.resolve(process.cwd(), "./.cdn-cache.json");
   }
 
   // 初始化cdn缓存
@@ -44,7 +60,10 @@ class CdnCache {
       this.cdnCache = JSON.parse(cdnCacheFileText);
     } catch (err) {
       console.log("cdn缓存文件不存在，创建缓存文件");
-      this.cdnCache = {};
+      this.cdnCache = {
+        packageDependencies: {},
+        cdnsUrl: {},
+      };
       await fs.writeFileSync(this.cdnCachePath, "", "utf-8");
     }
   }
@@ -55,7 +74,7 @@ class CdnCache {
    * @param version   版本
    */
   getCdnCache(packageName: string, version: string): CacheCellType[] | undefined {
-    return this.cdnCache[packageName]?.[version];
+    return this.cdnCache["cdnsUrl"][packageName]?.[version];
   }
   /**
    * 设置cdn缓存
@@ -64,10 +83,10 @@ class CdnCache {
    * @param urls  地址列表
    */
   setCdnCache(packageName: string, version: string, cdnData: CacheCellType[]) {
-    if (this.cdnCache[packageName]) {
-      this.cdnCache[packageName][version] = cdnData;
+    if (this.cdnCache["cdnsUrl"][packageName]) {
+      this.cdnCache["cdnsUrl"][packageName][version] = cdnData;
     } else {
-      this.cdnCache[packageName] = {
+      this.cdnCache["cdnsUrl"][packageName] = {
         [version]: cdnData,
       };
     }
@@ -77,6 +96,30 @@ class CdnCache {
    */
   async save() {
     await fs.writeFileSync(this.cdnCachePath, JSON.stringify(this.cdnCache), "utf-8");
+  }
+
+  /**
+   * 获取依赖包的package.json
+   * @param packageName  包名
+   * @param identification   版本号或者url
+   */
+  getPackageDependencies(packageName: string, identification: string): Dependencies | undefined {
+    return this.cdnCache["packageDependencies"][packageName]?.[identification];
+  }
+  /**
+   * 设置依赖包的package.json
+   * @param packageName  包名
+   * @param identification    版本号或者url
+   * @param dependencies  依赖
+   */
+  setPackageDependencies(packageName: string, identification: string, dependencies: Dependencies) {
+    if (this.cdnCache["packageDependencies"][packageName]) {
+      this.cdnCache["packageDependencies"][packageName][identification] = dependencies;
+    } else {
+      this.cdnCache["packageDependencies"][packageName] = {
+        [identification]: dependencies,
+      };
+    }
   }
 }
