+++ packages/vite-add-cdn-script/lib/utils/cdnProcess/index.ts
@@ -1,1 +1,2 @@
 export * from "./lib";
+export * from "./findUrls";
