+++ packages/vite-add-cdn-script/lib/utils/consoleManage.ts
@@ -1,4 +1,4 @@
-import { libName } from "../main";
+import { libName } from "../config";
 
 // 打印类型
 export enum ConsoleType {
@@ -48,6 +48,22 @@ export class ConsoleManage {
     console.log("");
     this.logList.forEach((item) => {
       console[item.type](`${libName} ${item.message}`);
+    });
+  }
+
+  public addMessageList(
+    addType: `${ConsoleType}`,
+    messages:
+      | string[]
+      | {
+          toString: () => string;
+        }[],
+  ) {
+    messages.forEach((item) => {
+      this.logList.push({
+        type: addType,
+        message: item.toString(),
+      });
     });
   }
