+++ packages/vite-add-cdn-script/test/consoleManage.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, jest, beforeEach, afterEach } from "@jest/globals";
 import { ConsoleManage } from "../lib/utils";
-import { libName } from "../lib/main";
+import { libName } from "../lib/config";
 
 describe("consoleManage", () => {
   const consoleManage = new ConsoleManage();
